Full breakout of different bot sections
Delegate different aspects to different modules for easier expansion.

Databases are handled by db, RCON client is hll_rcon_client, etc.
